stub for deleted movies
-обработка ошибки не существующего id фильма при открывании
-закрывание модалки тремя способами
