+++ src/js/modal.js
@@ -6,9 +6,13 @@ const refs = getRefs();
 refs.closeInfoBtn.addEventListener('click', onCloseInfoFilm);
 function onCloseInfoFilm(ev) {
     ev.stopPropagation();
-    onToggleClass()
-    refs.infoFilmContainer.innerHTML = '';
-    document.removeEventListener('keydown', onEventKey);
+    if (ev.code === 'Escape'
+            || ev.target.classList.contains('backdrop')
+            || ev.target.classList.contains('icon')) {
+        onToggleClass()
+        refs.infoFilmContainer.innerHTML = '';
+        document.removeEventListener('keydown', onEventKey);
+    }
 }
 
 refs.mainContainer.addEventListener('click', onOpenInfoFilm);
@@ -20,7 +24,7 @@ function onOpenInfoFilm(ev) {
     }
 }
 
-export default function onToggleClass() {
+function onToggleClass() {
     refs.infoFilmIsOpen.classList.toggle('backdrop--is-hidden');
     refs.bodyEl.classList.toggle('toggle_scroll');
 }
